Desafio: Aplicando o conhecimento em REST para Cidade.

+++ src/main/java/com/algaworks/algafoodapi/infrastructure/repository/CidadeRepositoryImpl.java
@@ -3,9 +3,12 @@ package com.algaworks.algafoodapi.infrastructure.repository;
 import com.algaworks.algafoodapi.domain.model.Cidade;
 import com.algaworks.algafoodapi.domain.repository.CidadeRepository;
 import java.util.List;
+import java.util.Objects;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
+
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,9 +41,12 @@ public class CidadeRepositoryImpl implements CidadeRepository
 
 	@Transactional
 	@Override
-	public void remover(Cidade cidade)
+	public void remover(Long id)
 	{
-		cidade = buscarPorId(cidade.getId());
+		Cidade cidade = buscarPorId(id);
+		if (Objects.isNull(cidade)){
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(cidade);
 	}
 }
